Fix migrations for electron
