+++ vite.config.ts
@@ -10,11 +10,50 @@ import path from 'node:path';
 import autoprefixer from 'autoprefixer';
 import VueI18nPlugin from '@intlify/unplugin-vue-i18n/vite'
 import electron from 'vite-plugin-electron'
+import { copyFileSync, existsSync, mkdirSync } from 'fs'
 
 
 
 
 const __dirname = dirname(fileURLToPath(import.meta.url))
+
+// Custom plugin to copy migrations and shared files
+const copyMigrationsPlugin = () => {
+  return {
+    name: 'copy-migrations',
+    closeBundle() {
+      // Copy migrations file
+      const migrationsSourcePath = path.resolve(__dirname, 'src/electron/migrations.js')
+      const migrationsTargetPath = path.resolve(__dirname, 'dist_electron/migrations.js')
+      
+      if (existsSync(migrationsSourcePath)) {
+        // Ensure target directory exists
+        const targetDir = path.dirname(migrationsTargetPath)
+        if (!existsSync(targetDir)) {
+          mkdirSync(targetDir, { recursive: true })
+        }
+        
+        copyFileSync(migrationsSourcePath, migrationsTargetPath)
+        console.log('✅ Migrations file copied to dist_electron/')
+      }
+
+      // Copy shared-utils file
+      const sharedSourcePath = path.resolve(__dirname, 'src/modules/shared-utils.js')
+      const sharedTargetPath = path.resolve(__dirname, 'dist_electron/shared-utils.js')
+      
+      if (existsSync(sharedSourcePath)) {
+        // Ensure target directory exists
+        const sharedTargetDir = path.dirname(sharedTargetPath)
+        if (!existsSync(sharedTargetDir)) {
+          mkdirSync(sharedTargetDir, { recursive: true })
+        }
+        
+        copyFileSync(sharedSourcePath, sharedTargetPath)
+        console.log('✅ Shared utils file copied to dist_electron/')
+      }
+    }
+  }
+}
 
 export default defineConfig(({ mode }) => {
   const isElectron = mode === 'electron'
@@ -60,10 +99,15 @@ export default defineConfig(({ mode }) => {
               include: [path.resolve(__dirname, 'src/locales/**')],
               strictMessage: false
             }),
+            copyMigrationsPlugin(),
           ],
           build: {
             outDir: 'dist_electron', // output folder for electron files,
             rollupOptions: {
+              input: {
+                background: "src/electron/background.js",
+                preload: "src/electron/preload.js"
+              },
               external: ['open', 'fluent-ffmpeg', 'ffmpeg-static', 'ffprobe-static']
             },
           },
